feature : Wrap array to ArrayProp
if you modify the array returned through __get with the array operator, an error occurs. (like "[]=" operator)

To avoid this,automatic  converted the array to an ArrayProp object.
You can use it as you would use a regular array.

You can cast it to a real array through the getAll() method.

----------------------------------

The reason of ArrayProp not extending the \ArrayObject is  is because it doesn't store an array.
Each time request a value, get value from origin objects or class.